Update logbin_expected.py
count.sum datatype is float64 not int64

+++ cooltools/cli/logbin_expected.py
@@ -135,7 +135,7 @@ def logbin_expected(
         "region": np.str,
         "diag": np.int64,
         "n_valid": np.int64,
-        "count.sum": np.int64,
+        "count.sum": np.float64,
         exp_summary_name: np.float64,
     }
